Move oauth setting resource to configuration package
Move SendConfigurationPatch() to configuration package
