+++ pkg/artifactory/resource/configuration/resource_artifactory_general_security.go
@@ -71,7 +71,7 @@ func resourceGeneralSecurityUpdate(ctx context.Context, d *schema.ResourceData, 
 		return diag.Errorf("failed to marshal general security settings during Update")
 	}
 
-	err = utils.SendConfigurationPatch(content, m)
+	err = SendConfigurationPatch(content, m)
 	if err != nil {
 		return diag.Errorf("failed to send PATCH request to Artifactory during Update")
 	}
@@ -87,7 +87,7 @@ security:
   anonAccessEnabled: false
 `
 
-	err := utils.SendConfigurationPatch([]byte(content), m)
+	err := SendConfigurationPatch([]byte(content), m)
 	if err != nil {
 		return diag.Errorf("failed to send PATCH request to Artifactory during Delete")
 	}
