+++ pkg/artifactory/resource/configuration/resource_artifactory_oauth_settings_test.go
@@ -1,4 +1,4 @@
-package security_test
+package configuration_test
 
 import (
 	"fmt"
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/acctest"
-	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/security"
+	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/configuration"
 )
 
 const OauthSettingsTemplateFull = `
@@ -108,7 +108,7 @@ func testAccOauthSettingsDestroy(id string) func(*terraform.State) error {
 		if !ok {
 			return fmt.Errorf("error: resource id [%s] not found", id)
 		}
-		oauthSettings := security.OauthSettings{}
+		oauthSettings := configuration.OauthSettings{}
 		_, err := client.R().SetResult(&oauthSettings).Get("artifactory/api/oauth")
 		if err != nil {
 			return fmt.Errorf("error: failed to retrieve data from <base_url>/artifactory/api/oauth during Read")
